+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add a new `Button` component.
+
 ## [1.12.0] - 2019-02-01
 
 Features: